requirements.txt
openai>=1.0.0
phoenix-evals>=0.4.0
phoenix>=0.4.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
gradio>=3.50.0
llama-parse>=0.3.0
nest-asyncio>=1.5.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
httpx>=0.24.0
pydantic>=2.0.0
